About
The technique exploits certain parameters which will aid to detect surface targets or obstacles such as a rock or a mine. In general machine learning, I give the date set and tell to the system to predict the data and show me the result whether it is a rock or mine. this technology is used in sonar. this is my first machine learning and I used a technology Jupiter notebook.